requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
apache-libcloud==3.8.0
    # via
    #   -r requirements.in
    #   django-storages
asgiref==3.7.2
    # via
    #   daphne
    #   django
    #   django-countries
asn1crypto==1.5.1
    # via webauthn
async-timeout==4.0.2
    # via redis
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
azure-core==1.24.2
    # via
    #   azure-storage-blob
    #   django-storages
azure-storage-blob==12.13.0
    # via
    #   -r requirements.in
    #   django-storages
bcrypt==3.2.2
    # via
    #   -r requirements.in
    #   paramiko
boto3==1.34.39
    # via
    #   -r requirements.in
    #   django-storages
botocore==1.34.39
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.3
    # via google-auth
cbor2==5.6.5
    # via webauthn
certifi==2024.7.4
    # via
    #   msrest
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   -r requirements.in
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.11
    # via requests
constantly==23.10.4
    # via twisted
cryptography==44.0.2
    # via
    #   autobahn
    #   azure-storage-blob
    #   paramiko
    #   pyopenssl
    #   service-identity
    #   webauthn
daphne==4.1.2
    # via -r requirements.in
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.21
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-anymail
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
    #   sentry-sdk
django-anymail==12.0
    # via -r requirements.in
django-cors-headers==3.10.1
    # via -r requirements.in
django-countries==7.6.1
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-storages[azure,boto3,google,libcloud,sftp]==1.14.2
    # via -r requirements.in
django-timezone-field==7.1
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r requirements.in
duo-client==4.2.3
    # via -r requirements.in
google-api-core==2.18.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.14.0
    # via
    #   -r requirements.in
    #   django-storages
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via google-cloud-storage
googleapis-common-protos==1.63.0
    # via google-api-core
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.7
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-spectacular
isodate==0.6.1
    # via msrest
jmespath==0.9.5
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
markdown==3.5.2
    # via -r requirements.in
more-itertools==8.2.0
    # via -r requirements.in
msrest==0.6.21
    # via azure-storage-blob
ntplib==0.3.4
    # via -r requirements.in
oauthlib==3.1.0
    # via requests-oauthlib
paramiko==3.4.0
    # via
    #   -r requirements.in
    #   django-storages
pillow==10.3.0
    # via -r requirements.in
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg[binary]==3.1.18
    # via -r requirements.in
psycopg-binary==3.1.18
    # via psycopg
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
    #   service-identity
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   service-identity
pycparser==2.21
    # via cffi
pycryptodomex==3.19.1
    # via -r requirements.in
pynacl==1.5.0
    # via
    #   -r requirements.in
    #   paramiko
pyopenssl==25.0.0
    # via
    #   twisted
    #   webauthn
pyotp==2.3.0
    # via -r requirements.in
python-dateutil==2.8.0
    # via
    #   -r requirements.in
    #   botocore
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   drf-spectacular
redis==4.5.4
    # via django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   apache-libcloud
    #   azure-core
    #   django-anymail
    #   google-api-core
    #   google-cloud-storage
    #   msrest
    #   requests-oauthlib
    #   yubico-client
requests-oauthlib==1.3.1
    # via msrest
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
s3transfer==0.10.0
    # via boto3
scrypt==0.8.27
    # via -r requirements.in
sentry-sdk[django]==2.16.0
    # via -r requirements.in
service-identity==24.1.0
    # via twisted
six==1.16.0
    # via
    #   azure-core
    #   duo-client
    #   isodate
    #   python-dateutil
sqlparse==0.5.0
    # via django
toml==0.10.2
    # via -r requirements.in
twisted[tls]==24.7.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.12.2
    # via
    #   azure-core
    #   dj-database-url
    #   django-countries
    #   psycopg
    #   pyopenssl
    #   referencing
    #   twisted
tzdata==2023.4
    # via -r requirements.in
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.19
    # via
    #   botocore
    #   django-anymail
    #   requests
    #   sentry-sdk
webauthn==2.5.2
    # via -r requirements.in
yubico-client==1.12.0
    # via -r requirements.in
zope-interface==7.1.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
